move addCompletedTourStep into app.tsx to prevent the tour from re-opening without refresh
